Update patch set 3
Patch Set 3: Code-Review+1

yep that is exactly what i was hoping to see
this looks good to me.

it would be interestign is we notice any reduction in the time it takes test to run. we might not but if we had slow start up of service or intermittent failure it might. we also have considered enable the RNG by default so if we see an overall imporvemtne in stablity of execution time that might be a useful factor to condier for nova in the future.

Patch-set: 3
Reviewer: Gerrit User 11604 <11604@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1