Update patch set 3
Patch Set 3:

There is no "python" binary on Centos8 afaict. Most distros won't symlink python3 to python because they think that this will confuse users who expect python always to be python2. So we will need to move to call python3 everywhere instead. This should always be installed in a devstack deployment, if this is missing somewhere, please file a bug and better yet fix it.

Patch-set: 3