Update patch set 1
Patch Set 1:

> I don't yet understand how or why the settings are not making it to
 > the conf file. Going to dig further.

OK, so I'm guessing we don't get the debug info bc it's failing on the old side of grenade, before we get this new devstack change. I'm gonna try proposing a DNM nova change to Depends-On this so we can get a failure with this change being used on the old side of grenade.

Patch-set: 1