Update patch set 5
Patch Set 5:

Python 3.6 is perfectly working on CentOS 7 and it doesn't requires any special trick to install it outher than executing:

 sudo yum install python3

This change is unlocking us when using python3 on CentOS 7. Because we are dropping support for Python 2 in many python projects, we should give high priority to this change.

Patch-set: 5