WIP: add get-os-environment task to tox playbook
I'm trying to get the python-cinderclient functional tests to run
using devstack-tox-functional as a base instead of our own local
playbooks (which pretty much just repeat the devstack-tox-functional
playbook defined in this repo).  The jobs aren't working, I think
because the get-os-environment task is missing.  So this patch adds
it.  Let's see iff that fixes https://review.opendev.org/#/c/691467/.

Change-Id: If64d8d39c7b26928b9de954cee71cb083c0f2416