Update patch set 3
Patch Set 3: Code-Review+2

Happy with this, but note that the fedora-latest tag is liable to change whenever we can get the latest happening.

Fedora 31 has hit a snag in that it changed the package repository format, and we have issues building in a chroot because of it.  We need to deploy more updated Bionic-based image build hosts which is taking some time to get sorted out.

Modulo that LGTM

Patch-set: 3
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2