Update patch set 3
Patch Set 3:

> clearly this is part of a larger stack for neutron, but for those
 > of us not following that could you provide a little more info why
 > we should add this in here by default?

Sure, I will comment here and then update the commit message as well.

We've started moving the code from networking-ovn to neutron as part of [0].

To run the OVN functional tests along with the neutron ones we need to install OVN when configuring the functional tests environment (see [1]) and for that we need this change here (an alternative would be to compile OVN from source but, that takes time).

Makes sense ?

[0] https://blueprints.launchpad.net/neutron/+spec/neutron-ovn-merge
[1] https://review.opendev.org/#/c/697440/

Patch-set: 3
Label: Verified=0