Update patch set 1
Patch Set 1:

> seems I haven't been clear enough. I'd drop both unstack.sh and clean.sh, just dump your instance and install a fresh one when you are done.
> 
> Anything else would require someone to set up a job that verifies that a stack/clean/stack cycle actually does work properly.

Is this just because of the lack of test coverage?

IMHO I believe unstack.sh or cleanup.sh is useful for local developments, so I would like to have clean.sh at least.

I am not sure what is a common pattern in development. I usually develop a patch in a different directory from /opt/stack. I re-run stack.sh (and unstack.sh) when I test the patch including the setup of OpenStack. If we only supports stack.sh (without unstack.sh or clean.sh), we always need to prepare a different machine and configure an access to the development machine.

I also usually use a baremetal machine (with something ironic or ubuntu maas). Deployments of baremetal machines tend to be longer compared to VMs in general, If we need to prepare a fresh machi, it means it forces me to wait longer.

Patch-set: 1