Update patch set 1
Patch Set 1:

Lack of test coverage combined with lack of developers and reviewers, so IMO reducing scope where possible is a good thing. I do agree that that local development workflow is useful, so if someone cares enough in order to actually implement the testing, that would seem a good enough criterion to me to keep it.

Patch-set: 1