Update patch set 4
Patch Set 4:

> The quoted change actually states that packages are not enough and
 > OVN must be compiled regardless - I might be wrong but it sounds
 > like it contradicts the usefulness of what this change proposes.

Sorry it was installing it from packages until the patch-set #11 [0] but then the PPA we were using (because we need a newer version than the ones that are in the normal repository) had a version conflict and we decided that for now we will go with compiling from source.

But, installing OVN from packages is still our goal.

[0] https://review.opendev.org/#/c/697440/11/tools/configure_for_func_testing.sh

Patch-set: 4