Update patch set 4
Patch Set 4:

tag 18.0.0 is not working due to --black_regex not available. Trying tag 19.

This is other issue where job inventory is in master branch where roles of run-tempest is defined as per the master code which is not compatible with old Tempest tag. Not sure how to fix that.

Patch-set: 4
Label: Verified=0