Update patch set 6
Patch Set 6:

This is a very important patch and I think it is taking too much time to get merged. I would suggest for the next time to try have all this changes reviewed in separated mini-patches, all of them as mucha as important to have something incrementally advancing. Thank you very much for working on this.

Patch-set: 6