Update patch set 10
Patch Set 10:

I do wonder if we really need it - it seems requirements' bot is just too eager to break u-c - since this is closing em it might be better to pin the versions by ensuring u-c are not too demanding. OTOH, it feels wrong to keep 3.5 that long in requirements but it's essentially how zipp issue was solved just today. Just food for thought.

Patch-set: 10