Update patch set 4
Patch Set 4:

> I almost have f30 nodes deployed ... next week.  then i think we can test this

Just for everyone's FYI ... I should note this hit a rather major snag -- the RPM package format changed and we can not build minimal Feodra images on < bionic because the rpm we use on the host doesn't understand how to uncompress the bootstrap packages.

That has not been as easy to work around as you might think.  I *can* just create a bionic builder and have it work ... but nodepool and zuul are all in the middle of containerisation efforts.  the idea is to run them out of containers rather than our old method of pip installing them globally.  this is a much bigger thing to deploy into production and has led to the inordinate hold-up of getting an environment to actually test this.

Patch-set: 4
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0