Update patch set 2
Patch Set 2:

Looks like the iscsi-initiator-utils bug [1][2] is still present even with the patched version from iwienand [3] being installed on the nodes.

The actual patch landed in 2.0.877 and 2.1.0 is now available from rawhide so I might rebuild that version for 30 to see if that resolved the issue.

$ git tag --contains baa0cb45cfcf10a81283c191b0b236cd1a2f66ee
2.0.877
2.0.878
2.1.0

[1] http://zuul.opendev.org/t/openstack/build/55fd498e721a4043bcbd384427c9e048/log/controller/logs/screen-n-cpu.txt#35587
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1676365
[3] https://copr.fedorainfracloud.org/coprs/iwienand/iscsi-initiator-utils/

Patch-set: 2
Label: Verified=0