Update patch set 2
Patch Set 2: Code-Review-1

Adding link to release where it was removed can help here. grenade job is happy here but as frickler mentioned it will be good if you can push Depends-On patch for nova, neutron, swift etc so that devstack not break any of their job relying on g-reg. 

one more question: I searched for g-reg usage and many projects enabling it in their devstack plugin or jobs. can we cleanup those first to be clear that nobody required these. I am -1 for this mainly.

- http://codesearch.openstack.org/?q=g-reg&i=nope&files=&repos=

Patch-set: 2
Label: Code-Review=-1