Update patch set 2
Patch Set 2:

> > Adding link to release where it was removed can help here.
 > grenade
 > > job is happy here but as frickler mentioned it will be good if
 > you
 > > can push Depends-On patch for nova, neutron, swift etc so that
 > > devstack not break any of their job relying on g-reg.
 > >
 > > one more question: I searched for g-reg usage and many projects
 > > enabling it in their devstack plugin or jobs. can we cleanup
 > those
 > > first to be clear that nobody required these. I am -1 for this
 > > mainly.
 > >
 > > - http://codesearch.openstack.org/?q=g-reg&i=nope&files=&repos=
 > 
 > Just asking out of curiosity;
 > Isn't it the responsibility of dependent projects when some
 > deprecation is announced then they should get rid of the deprecated
 > part, or it is the responsibility of the host project to keep track
 > who all are using that deprecated part and make sure those are not
 > affected when deprecated part is removed.

True when it comes on projects side but for QA tooling it is not easy. We live in a complex world of not breaking anyone for any reason :).

I am not worried about the users of g-reg which should care by themself as you mentioned but I am worried about users of "devstack setup of g-reg" by removing that we have to make sure no one replying on that, see this as example which hanging so long - https://review.opendev.org/#/q/topic:volume-default-testing-v3+(status:open+OR+status:merged) 

I am ok if you notify them via patch up or ML and we can leave inactive projects if not responding on pathces or ML.

Patch-set: 2