Add support for IPv6 tunnel endpoints
Currently, neutron tunnel endpoints must be IPv4 addresses,
i.e. $HOST_IP, although IPv6 endpoints are supported by most
drivers.

Create a TUNNEL_IP_VERSION variable to choose which host IP
to use, either HOST_IP or HOST_IPV6, and configure it in the
OVS and Linuxbridge agent driver files. Default is IPv4, but
the goal is to eventually setup the "IPv6-only" jobs to use
it to truly have only IPv6 addresses being used. It can be
over-ridden by specifying TUNNEL_ENDPOINT_IP accordingly.

This behaves similar to the SERVICE_IP_VERSION option, which
can either be set to 4 or 6, but not 4+6 - the tunnel overhead
should be consistent on all systems in order not to have MTU
issues.

Change-Id: I6034278dfc17b55d7863bc4db541bbdaa983a686