Install swift and cinder client from pip if needed
We have logic to install swiftclient if swift is enabled because
glance uses swift by default if it's enabled. However, if swiftclient
isn't in the list of libs to install from source, we no-op. This was
fine when we were accidentally getting python-swiftclient from
the test-requirements of glance-store, but it's an actual need, so
we need to also fall back to installing from pip. Same story
with cinderclient which is also pulled in originally from glance_store
test-requirements.

Change-Id: I658d190557a993f3cfe6ed55e5d2159f47f7d35e