Update patch set 1
Patch Set 1:

> could use making it green; also, let's maybe switch nomenclature to
 > include fedora version rather than claiming latest

The issue with changing the naming now is that these jobs are referenced by other projects (such as Nova) and so changing that here is going to break things. Would it be possible to delay this renaming until early in the V cycle when the gate is less full and it would be easier to get things through?

Patch-set: 1
Reviewer: Gerrit User 10135 <10135@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0