Update patch set 1
Patch Set 1: Workflow-1

I'm still not sure about this but wanted to post something as Radosław asked about it again in the f31 series.

Given the 6 month lifecycle of Fedora and the number of projects currently using the fedora-latest nodeset or jobs directly is this worth the churn? I'll take this to the ML now as well.

Patch-set: 1
Reviewer: Gerrit User 10135 <10135@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=-1