Update patch set 10
Patch Set 10:

Most test instances seems to be rather limited on IOPS, so avoiding an intermediate copy on disk we seem good. Not sure if we'd have enough space on ramdisk instead, using 2G of our 8G total might lead to other issues. In the end, using a blown up image like this in CI might not be feasible at all.

Patch-set: 10
Label: Verified=0