Update patch set 1
Patch Set 1:

Hi, based on bug report and comment above the fix, I guess the claim that stdin was parsed as utf-8 was false. I would advise that you check locally if you use a UTF-8 locale. If not, then we could think about smoothing out the experience for non-UTF-8 locales; if yes, then we might have to check where that surrogate happens as it should not.

Patch-set: 1