Update patch set 1
Patch Set 1:

> this is really cumbersome, I think we would be better with a
 > per-release diff in tempest itself (that way it would always track
 > the current per request)

yeah extensions are not well managed so doing it manually is the way. It is default to 'all' on Temepst side and manage per release ext on devstack side. This make sure we do not accidently run the extensions tests not present in that stable branch.

Patch-set: 1
Reviewer: Gerrit User 8556 <8556@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0