Use native sgabios.bin
In recent distrubtions, for example Ubuntu 20.04, the sgabios.bin
ROM is provided directly by qemu-system-data as an actual file
under /usr/share/qemu and it conflicts with the one provided by
sgabios.
We need to copy the one under /usr/share/misc only if the other
one is a symlink, meaning it's provided by sgabios.

Change-Id: Ib5f23dbd8839a0927418692054f4ed4abd76babc