Use stable constraint for older Tempest
stable/train|stein are supported stable and so Tempest
master version is used for their testing. But we still have
python2 jobs on their gate and they are tested with older compatible
Tempest(23.0.0). Tempest 23.0.0 is not compatible with master
constraint anymore
- I2d58d1b524764ea1d46d6ab3a382783f6a002b34

and started failing the py2 jobs with older Tempest on
stable/train|stein gate.
- https://stash.dellemc-community.org/logs/dellemc-poweredge-R730-R630-tempest-dsvm-ironic-idrac/2020-05-19:10:37/727906/5/logs/

To use the compatible constraint, we need to take stable constraint
for older Tempest and master constraint for master Tempest.

Change-Id: I11123a34297d90226077b9f559297af627ac6e38