Handle uwsgi install for openSUSE
Add packages for openSUSE distribution for installation of uwsgi.

For stein add in addition:
Adapt PYTHON_PATH and add /usr/local/bin to PATH. It is not in PATH
because we don't pick up /etc/profile where it is usually added. This is
needed to fix the setup after the pip/virtualenv changes.
Also install python2 version of uwsgi module in addtion to 3.

The depended on patch is to set PATH for temptest for the same reason.

Depends-On: https://review.opendev.org/#/c/739778/
Co-Authored-By: Jens Harbott <j.harbott@x-ion.de>
Co-Authored-By: Jan Zerebecki <jzerebecki@suse.com>
Change-Id: I2ce8959460a79b6472bd9dd59edd7d94eccbacf5
(cherry-picked from commit 10c3ffd26557f1921841e456b003a77f6fb11948)