Update patch set 2
Patch Set 2: Code-Review+1

> True when it comes on projects side but for QA tooling it is not
 > easy. We live in a complex world of not breaking anyone for any
 > reason :).
 > 
 > I am not worried about the users of g-reg which should care by
 > themself as you mentioned but I am worried about users of "devstack
 > setup of g-reg" by removing that we have to make sure no one
 > replying on that, see this as example which hanging so long -
 > https://review.opendev.org/#/q/topic:volume-default-testing-v3+(status:open+OR+status:merged)
 > 
 > I am ok if you notify them via patch up or ML and we can leave
 > inactive projects if not responding on pathces or ML.

Gmann, glance-registry has been neutered in-tree such that it cannot do any useful work. That has already merged, and hasn't broken other projects, which means it is clearly not a dependency of any other project. See here:

https://review.opendev.org/#/c/734215

They are keeping this code in tree for the moment just to avoid breaking devstack, but that obviously needs to be removed, along with the devstack code for g-reg now that it has been shown not to break anyone.

The devstack glance module needs other work in order to fix some other features so we can test them (both Nova and Glance teams need this), and doing so will be much easier if we can remove the g-reg setup code from devstack first.

So, I think it has been shown that there are no dependencies on this, and grenade is happy. Are we good to move forward here?

Patch-set: 2
Label: Code-Review=+1