Fix for downloading uwsgi on centos in lib/apache
On Centos the lib/apache script in devstack downloads and compiles
uwsgi package. It uses pip-download to pull uwsgi tar file.
However, it does not return the full name of downloaded file so
right now the script looks for filename that starts with prefix
"uwsgi".
This method is case sensitive so after downloading file with name
uWSGI-2.0.19.tar.gz the script was unable to locate this file.

This change downloads the file to an empty directory and expect
it to be the only file there so there is no name vulnerability.

Change-Id: I68027c70ca8eb3059d14f3e35cc7d97941f5dc46
Closes-Bug: #1883897