Update patch set 3
Patch Set 3:

I think this should already fix the CentOS case by using the same installation way as Fedora: https://review.opendev.org/#/c/735558/

Anyway this patchset could have some kind of value for other distributions or when RPMs approach is not the preferred one in case of this problem hasn't been fixed yet.

Anyway in this moment we have another issue [1] with CentOS 8 job [2] that is not allowing it to work.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1847979
[2] https://zuul.opendev.org/t/openstack/builds?job_name=devstack-platform-centos-8&project=openstack/devstack

Patch-set: 3