Update patch set 2
Patch Set 2:

> tempest-full-py3 job is failing with below error;
 > 2020-06-17 05:58:03.395259 | controller | ++ lib/tempest:configure_tempest:155
 >        :   '[' '' = cirros-0.5.1-x86_64-disk ']'

This looks to me like we don't read even a single line containing an image, likely because osc is just returning something on stderr maybe?

Maybe slap a bare "openstack images list" in front of this so we can see what is happening? Does this happen in a local devstack too, or only in CI?

Patch-set: 2