Update patch set 2
Patch Set 2:

> Actually this is a while loop, condition to read images is at line
 > 160 [1]

Yeah, I know, a loop to read one line at a time. It seems to read one line and exit, with no image name/uuid on that line right?

 > The output of above line looks like;
 > openstack image list --property status=active | awk -F'|'
 > '!/^(+--)|ID|aki|ari/ { print $3,$2 }'
 > 
 > cirros-0.5.1-x86_64-disk   72aedb66-1aa0-46f6-a79d-d127042c1f94

I know it's supposed to look like that. I'm saying it clearly seems to not in this case. I'm wondering if the image import results in images not going to active quick enough when you use the glanceclient case, such that we race to here, do a list for active and get nothing.

Patch-set: 2