Update patch set 3
Patch Set 3:

> How about doing this special stuff in a glance devstack plugin like
 > neutron does for sophisticated features? That would avoid the
 > political discussion of OSC vs. glanceclient.

Just to be clear, we want the cirros image that is uploaded during the main devstack phase to be import-converted like it would in a real deployment before it gets used by the tempest tests. So, yes, we could just do this completely out of band to make sure that the basic feature works, and I guess we could override tempest config to use the *other* image that we imported instead of the one devstack created, but that would be kind of silly and obscure for debugging no?

What is the motivation behind requiring everything in devstack to be done by OSC? Is it just the clouds.yaml support that is the sticking point? What devstack or infra requirement brings that forth?

I would maintain the position that devstack serves not to be a pinnacle of deployment purity, nor a motivator of cross-project goals, but a developer tool for testing openstack. Has devstack changed scope such that it has its own motivations for making projects do certain things independent of what the users of those projects need or the developers of those projects provide?

Patch-set: 3