Update patch set 2
Patch Set 2:

> Why is the -W? The patch looks OK to me, does what we aimed it for:
 > tempest job picks up the process-stackviz role from devstack
 > instead of from tempest.
 > 
 > This patch needs to be merged to ussuri, train and stein to make
 > the gate pass in tempest master for the process-stackviz removal
 > patch [1]. Until that patch does not removes the process-stackviz
 > from master, this patch in devstack does not cause any problem as
 > zuul picks the process-stackviz role from devstack.
 > 
 > [1] https://review.opendev.org/#/c/737074/

OK, I see now, that the problem is, that the new role is not used from devstack until the old role is not removed from tempest.
So if I'm not mistaken, this will not solve the issue in old branches, where Tempest is pinned to 20.0.0 as it would need a new release (20.0.1?) without the process-stackviz role.

I know that Tempest is branchless, but in this case can we make an exception and create a new branch or somehow create a patch on top of 20.0.1 to not to include process-stackviz anymore?

Patch-set: 2