Update patch set 4
Patch Set 4:

because this remove the request for installing python package on Ubuntu I verified this also unlock a problem running stack.sh on a standard Ubuntu Focal that doesn't provide any source for 'python' package. So this change also fixes these other problem a part of probably others for other non-ubuntu platforms like CentOS 8. This is a strong +1 for me. Thank you Jens!

Patch-set: 4