Add ensure-pip to pre.yaml
After running Tempest, it tries to install stackviz but that fails
due to old version of pip (version 8), which cannot handle correctly
the opendev mirror and fails to fetch some packages. This bug
disappears by pip version 9.

Tempest has ensure-pip before the failing role, but that's too late
as pip is already installed and won't be upgraded.

If ensure-pip is added in devstack's pre.yaml, then it will install
the correct version of pip from the beginning.

Change-Id: I76726aee4b6be98123e55fea58cd6e816f3ceb24