Enable extend attached volume tests for Cinder generic NFS driver
Cinder has a feature 'extend attached volume' that the drivers might
implement. Such drivers should run the specific tests for this feature,
but the flag to enable those tests is disabled for generic NFS driver.

This patch enables the 'extend attached volume' tests for the generic
NFS driver, as it's reporting that it supports the feature.

The tests will fail until bug #1870367 is fixed. This patch is necessary
to validate the bugfix patches.

Change-Id: I27084ec6da76b9c43ea0899fc6f1e842286b3dd3
Related-Bug: #1870367