Update patch set 7
Patch Set 7:

Federico, yes I agree. It seems that the master branch is working perfectly fine now.

Andrii, I can see that both rocky and queens branches have the following check:

elif [[ $os_VENDOR == "Fedora" ]]; then
    install_package uwsgi uwsgi-plugin-python3

At the same time master branch solves it this way:
elif is_fedora; then   
    install_package uwsgi uwsgi-plugin-python3

It causes that other branches than master try to compile uwsgi manually on centos which will fail. This problem should be fixed on these branches and there is a following question for you Jens:

Should we apply this fix as a workaround to the problem or perhaps we should change this if statement on other branches so Centos wouldn't try to compile this package on its own.

Thank you all for the review on this change and for the discussion.

Patch-set: 7
Reviewer: Gerrit User 31956 <31956@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0