Update patch set 6
Patch Set 6:

(2 comments)

> I'm still uncomfortable with this because a lot of the different
 > image variant aren't being tested in the normal CI jobs. Can we
 > instead just modify what is needed for the nova job?

The reason this is a bigger change than it needs to be is because the specific osc incantation for each are just copy/pasted everywhere, which is really messy, IMHO. If I were to make the minimal change, I'd still be touching each one of those osc calls, which wouldn't be tested, and just further propagating the lack of DRYness that makes this so ugly.

This is tested for the primary deployment types and works, and if the others test their cases once, then going forward we'll have a solid re-usable image upload function that will help avoid this concern in the future.

Patch-set: 6
Label: Verified=0