Update patch set 1
Patch Set 1: Workflow-1

This is an untested stab at making Glance able to be deployed standalone separate from everything else, keeping the tls-proxy intact even still.

I'm stacking with this locally, but wanted to put this up for early eyes and to get a test run of the default config.

Patch-set: 1
Label: Workflow=-1