Update patch set 7
Patch Set 7: Code-Review+1

(2 comments)

I am ok to get these bits in and start testing the import workflow. Best is to have glance run in uwsgi but as dan/abhshek mentioned those are big changes on glance side and we have been discussing those for years and did not happen for whatever reason. I cannot say only manpower is the reason for that but let's not get into those things again and let's the feature in existing way.

Waiting for the nova job result on latest results.
later we can add the nova-ceph-multistore job (once dan finish that) here (as n-v first) to have this tested on devstack gate too.

Patch-set: 7
Label: Code-Review=+1