Update patch set 7
Patch Set 7:

> I am ok to get these bits in and start testing the import workflow.
 > Best is to have glance run in uwsgi but as dan/abhshek mentioned
 > those are big changes on glance side and we have been discussing
 > those for years and did not happen for whatever reason. I cannot
 > say only manpower is the reason for that but let's not get into
 > those things again and let's the feature in existing way.

Just to be clear, I don't think the desire is to change glance to work better under uwsgi. The changes to make it work better under uwsgi would be large and involve substantial/fundamental changes to glance, which is why they have not happened. AFAIK, the community's move to uwsgi is not a goal shared by the glance team, and for understandable reasons.

Point being, I don't think changes to glance for more native uwsgi behavior are on the horizon any time soon, if ever, and not just for manpower reasons.

 > Waiting for the nova job result on latest results.
 > later we can add the nova-ceph-multistore job (once dan finish
 > that) here (as n-v first) to have this tested on devstack gate too.

FWIW, the changes in this patch only affect the early setup and import of the image, and not the tempest run, which has been failing for nova, glance and policy reasons lately. However, I think the latest run here will have all the stars aligned.

Patch-set: 7
Label: Verified=0