Make ironic cross gating voting
The majority of ironic's CI is built upon devstack, and thus some changes
which may pass setting up a stock devstack, may still ultimately cripple
Ironic's CI. Mainly in terms of integration with other projects.

As such, it seems logical to make the CI job voting such that breaking
changes have visibility upfront.

Also re-names the job name to the more modern name as the
driver name the job mentions was explicitly removed some time
ago in favor of what is called driver composition.

This change came out of a discussion of Neutron changing the default to
OVN. Which while problematic for Ironic due to a lack of necessary DHCP
server capabilities, the topic of the neturon-legacy/neutron plugin came
up and Ironic, and at the time of this change Ironic uses the "neutron"
plugin. As such, abrupt removal of the plugin would be a fairly
catastropic change for Ironic, hence the need for cross-gating.

Depends-on: https://review.opendev.org/#/c/741254/
Change-Id: I5d6a4fccf37521f0104091ecb8055a08f760fa05