Fix glance image handling
Update the image upload command for kernels/ramdisks so partition
image based CI integration testing jobs, like those used by ironic,
are not failing due to only the partition image being uploaded without
a kernel and ramdisk, which causes code interpreting the image to
believe that the image is infact a whole-disk-image.

This is needed as a result of change:
https://review.opendev.org/#/c/735966

+ functions:upload_image:427 : '[' -n /opt/stack/devstack/files/images/cirros-0.5.1-x86_64-uec/cirros-0.5.1-x86_64-initrd ']'
++ functions:upload_image:428 : oscwrap --os-cloud=devstack-admin --os-region-name=RegionOne image create cirros-0.5.1-x86_64-uec-ramdisk hw_rng_model=virtio --public --container-format ari --disk-format ari

Where previously "--property" was prepended to the properties set,
however for partition images, it doesn't make sense to support
explicit property setting as the kernel and ramdisk id values are
linked to by the primary glance image object once uploaded.

Change-Id: I86ffa2d963decb8492927f3d67f1d58b35870aac